chore: Enable unconvert linter and fix reported issues (#182)

+++ context/context.go
@@ -119,7 +119,7 @@ func (context *AMFContext) TmsiAllocate() int32 {
 		return -1
 	}
 	logger.ContextLog.Infof("Allocate TMSI : %v", val)
-	return int32(val)
+	return val
 }
 
 func (context *AMFContext) AllocateAmfUeNgapID() (int64, error) {
